Move Effect struct to protocols
